// src/app/core/api/model/appAuditEventSummaryDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AppActorDto } from './appActorDto';


export interface AppAuditEventSummaryDto { 
    id: string;
    eventType: AppAuditEventSummaryDto.EventTypeEnum;
    actor?: AppActorDto;
    changeMetadata: object;
    /**
     * Set only for DEPLOY and ROLLBACK events
     */
    revisionNumber?: number;
    imageRef?: string;
    createdAt: string;
}
export namespace AppAuditEventSummaryDto {
    export const EventTypeEnum = {
        Deploy: 'deploy',
        Rollback: 'rollback',
        Scale: 'scale',
        ResourceUpdate: 'resource_update',
        Restart: 'restart',
        Start: 'start',
        Stop: 'stop',
        ConfigUpdate: 'config_update',
        Reconciled: 'reconciled',
        Created: 'created',
        Delete: 'delete'
    } as const;
    export type EventTypeEnum = typeof EventTypeEnum[keyof typeof EventTypeEnum];
}